feat: 상품 조회 추가
- 상세 조회, 목록 조회 추가
- 조회 시 좋아요 개수 및 사용자의 좋아요 여부 포함
- 단위 테스트 추가

+++ apps/commerce-api/src/test/java/com/loopers/domain/like/LikeServiceTest.java
@@ -63,7 +63,7 @@ public class LikeServiceTest {
             userRepository.save(user);
             productRepository.save(product);
 
-            int countBefore = likeRepository.countAll();
+            long countBefore = likeRepository.countAll();
 
             // act
             likeService.addLike(user.getId(), product.getId());
